# BackEnd/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TsunamiIT Manager Project API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/scrumteam/{idSCRUMTeam}/meeting": {
            "post": {
                "tags": [
                    "Meetings"
                ],
                "summary": "Create meeting for a scrum team",
                "operationId": "93ed0544b5cb8a44758d019aa34bc031",
                "parameters": [
                    {
                        "name": "idSCRUMTeam",
                        "in": "path",
                        "description": "Scrum team ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "meeting1"
                                    },
                                    "date": {
                                        "type": "string",
                                        "example": "2022-08-30 10:02:05"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Token Not found "
                    },
                    "500": {
                        "description": "Server Error  "
                    }
                }
            }
        },
        "/api/meeting/{idMeeting}/addCollaborator": {
            "post": {
                "tags": [
                    "Meetings"
                ],
                "summary": "add collaborator to meeting",
                "operationId": "3282c38fd7a7f15c3a64fc770b89039a",
                "parameters": [
                    {
                        "name": "idMeeting",
                        "in": "path",
                        "description": "meeting ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "idCollaborator"
                                ],
                                "properties": {
                                    "idCollaborator": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Token Not found "
                    },
                    "500": {
                        "description": "Server Error  "
                    }
                }
            }
        },
        "/api/scrumteam/{idSCRUMTeam}/meetings": {
            "get": {
                "tags": [
                    "Meetings"
                ],
                "summary": "Returns meetings by scrum team ",
                "operationId": "7658f6e765dd8d087bd396adf0ad1125",
                "parameters": [
                    {
                        "name": "idSCRUMTeam",
                        "in": "path",
                        "description": "Scrum team ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation !"
                    }
                }
            }
        },
        "/api/collaborator/{idCollaborator}/meetings": {
            "get": {
                "tags": [
                    "Meetings"
                ],
                "summary": "Returns meetings by collaborator ",
                "operationId": "9a66350d793164e2216a1e036e9157b4",
                "parameters": [
                    {
                        "name": "idCollaborator",
                        "in": "path",
                        "description": "Collaborator ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation !"
                    }
                }
            }
        },
        "/api/meeting/{idMeeting}/uninvited": {
            "get": {
                "tags": [
                    "Meetings"
                ],
                "summary": "Returns uninvited collaborators ",
                "operationId": "79fc068ebf2360f7154e3459f40f827c",
                "parameters": [
                    {
                        "name": "idMeeting",
                        "in": "path",
                        "description": "Meeting ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation !"
                    }
                }
            }
        },
        "/api/joinmeeting/{idJoinMeeting}/declareAbsence": {
            "put": {
                "tags": [
                    "Meetings"
                ],
                "summary": "declare absence",
                "operationId": "55b0b7be28a94cbd719443adbe92b89c",
                "parameters": [
                    {
                        "name": "idJoinMeeting",
                        "in": "path",
                        "description": "Join Meeting ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "joinmeeting not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/joinmeeting/{idJoinMeeting}/cancelInvite": {
            "delete": {
                "tags": [
                    "Meetings"
                ],
                "summary": "cancel invitation by deleting joinmeeting row",
                "operationId": "2c1db0f683920888ea5a13f10b2b4ad5",
                "parameters": [
                    {
                        "name": "idJoinMeeting",
                        "in": "path",
                        "description": "Join meeting ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "join meeting not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    }
}